requirements.txt
Wand==0.6.11
pandas==2.0.3
xlrd >= 2.0.1
Pillow>=7.1.2
openpyxl==3.1.2
esdk-obs-python>=3.23.9